refactor: move git module to lib/git/
