+++ src/context-check.ts
@@ -1,6 +1,6 @@
 import pc from 'picocolors';
 import { resolve } from 'path';
-import { cloneRepo, cleanupTempDir } from './git.ts';
+import { cloneRepo, cleanupTempDir } from './lib/git/index.ts';
 import { parseSource } from './source-parser.ts';
 import { discover, filterByType } from './context-discovery.ts';
 import { executeInstallPipeline } from './context-installer.ts';
