+++ src/lib/git/gitignore.test.ts
