+++ src/lib/git/gitignore.ts
