+++ src/lib/git/index.ts
@@ -1,0 +1,3 @@
+export { cloneRepo, cleanupTempDir, GitCloneError } from './git.ts';
+export { addToGitignore, removeFromGitignore, readManagedPaths } from './gitignore.ts';
+export { fetchRepoTree, type GitHubTreeEntry } from './github-trees.ts';
